requirements.txt
glob2==0.7
Flask==2.0.1
Flask_Cors==3.0.10
ipykernel==5.4.3
imageio==2.15.0
ipdb==0.13.9
ipython==8.0.1
ipython-genutils==0.2.0
jupyterlab==3.0.6
jupyter-client==6.1.11
jupyter-core==4.7.1
JLDracula==0.1.0
matplotlib==3.3.4
numpy==1.21.2
Pillow==9.0.1
scipy==1.6.3
scikit-learn==0.24.1
torchvision==0.11.3
tqdm==4.62.1
Sphinx==1.8.6
wrapt==1.12.1